requirements.txt
Django==2.1.5
djangorestframework==3.10.3
django-filter==2.2.0
django-cors-headers==3.1.0
gunicorn==19.9.0
whitenoise==5.0.1
